test(agent): split L2 turn-aware tests into frame_ops_l2_test.go (round-2 F3)
frame_ops_test.go grew to ~6050 lines after L2 added ~1846 lines of new
tests at the tail. Body-of-file additions hurt blame ergonomics (every
new ApplyFrameEvent_TurnAwareProxyDetach row sits past the 5000-line
mark) and IDE jump (test-explorer entries land far from the helpers
they exercise).

Pure file move: extract every L2 test (TestSubagentRefMatches_TurnAware
through TestApplyFrameEvent_TurnAwareProxyDetach) and the L2-only
helpers (turnAwareEnvAlive / turnAwareEnvNoParent / seedProxyRef /
rawTurn) into a new sibling test file in the same package. Zero logic
change; the new file shares the existing test fixtures
(newProxyTestModule, seedFrame, fakes) via package-level scope.

frame_ops_test.go drops `sync` from its import set since the only
goroutine-using tests moved with the L2 block; frame_ops_l2_test.go's
imports (encoding/json, sync, testing, agentpkg, store) are the minimal
set actually used.

go test ./... + go test -race ./... + go vet ./... all clean.
